fix: rename kebab-case files to camelCase for valid JS exports
- Renamed pane-menu.ts to paneMenu.ts
- Renamed pane-store.ts to paneStore.ts
- Updated imports in playground App.tsx
- Fixed codegen exports to use valid JavaScript identifiers
