+++ packages/playground/src/App.tsx
@@ -5,7 +5,7 @@ import { Label } from '@openagentsinc/ui/web/components/label'
 import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@openagentsinc/ui/web/components/card'
 import { PaneManager } from '@openagentsinc/ui/web/components/pane/pane-manager'
 import { Hotbar } from '@openagentsinc/ui/web/components/hotbar/hotbar'
-import { usePaneStore } from '@openagentsinc/ui/web/stores/pane-store'
+import { usePaneStore } from '@openagentsinc/ui/web/stores/paneStore'
 import { Code, FileText, Settings, User, Bot } from 'lucide-react'
 import type { Pane } from '@openagentsinc/ui/core/types/pane'
 import './App.css'
