+++ packages/ui/src/index.ts
@@ -1,5 +1,14 @@
+
 export * as pane from "./core/types/pane.js"
-export * as paneMenu from "./core/types/pane-menu.js"
+
+
+export * as paneMenu from "./core/types/paneMenu.js"
+
+
 export * as cn from "./core/utils/cn.js"
+
+
 export * as os from "./core/utils/os.js"
-export * as paneStore from "./web/stores/pane-store.js"
+
+
+export * as paneStore from "./web/stores/paneStore.js"
